requirements.txt
# Core
tensorflow>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0
imagehash>=4.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
joblib>=1.1.0

# Testing
pytest>=6.2.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0